Изменение цветовой схемы

+++ src/ru/smak/math/Fractal.java
@@ -3,5 +3,5 @@ package ru.smak.math;
 import ru.smak.math.Complex;
 
 public interface Fractal {
-    float isInSet(Complex c);
+    double isInSet(Complex c);
 }
